Bump xterm.js to v3.12.0 and bundle its dist
in order to pick up latest bug fixes.

Direct inclusion of xterm.js dist is being done since the latest versions of xterm.js are not available via Bower (xterm.js-next is not being maintained).

Note:  this change results in invalid meta in bower.json (`The "main" field has to contain only 1 file per filetype...`), but the net result is only a warning in Bower's output when installing.  This is a small concession in order to easily keep xterm.js updated.